refactor: colocate current node updater
